# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Atomic Identity Habits API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ─── IDENTITIES ─────────────────────────────────────────────────────────────
...
  /ai/status:
    get:
      operationId: getAiStatus
      tags: [ai]
      summary: Check AI / Ollama availability
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiStatus"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a chat message to the AI coach
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatReply"
  /ai/confirm-action:
    post:
      operationId: confirmAiAction
      tags: [ai]
      summary: Confirm or reject an AI proposed action
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiActionConfirmInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiActionResult"
  /ai/memories:
    get:
      operationId: listAiMemories
      tags: [ai]
      summary: List AI memories
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMemory"
    post:
      operationId: createAiMemory
      tags: [ai]
      summary: Create an AI memory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMemoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMemory"
  /ai/memories/{id}:
    delete:
      operationId: deleteAiMemory
      tags: [ai]
      summary: Delete an AI memory
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /ai/conversations:
    get:
      operationId: listAiConversations
      tags: [ai]
      summary: List AI conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiConversation"
    post:
      operationId: createAiConversation
      tags: [ai]
      summary: Create a new AI conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConversation"
  /ai/conversations/{id}/messages:
    get:
      operationId: listAiMessages
      tags: [ai]
      summary: List messages in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"

  # ─── EXPORT ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string

    # ─── IDENTITY ────────────────────────────────────────────────────────────
...
    AiStatus:
      type: object
      required: [available, model]
      properties:
        available:
          type: boolean
        model:
          type: string
        message:
          type: ["string", "null"]
    ChatMessageInput:
      type: object
      required: [message]
      properties:
        conversationId:
          type: ["integer", "null"]
        message:
          type: string
        contextChips:
          type: array
          items:
            type: string
    ChatReply:
      type: object
      required: [message, conversationId]
      properties:
        message:
          type: string
        conversationId:
          type: integer
        proposedAction:
          type: ["object", "null"]
          properties:
            type:
              type: string
            payload:
              type: object
    AiActionConfirmInput:
      type: object
      required: [proposedActionType, payload, confirmed]
      properties:
        proposedActionType:
          type: string
        payload:
          type: object
        confirmed:
          type: boolean
    AiActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: ["object", "null"]
    AiMemory:
      type: object
      required: [id, userId, content, sourceType, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        sourceType:
          type: string
          enum: [JOURNAL, HABIT, USER_NOTE, AI_SUMMARY, REFLECTION]
        tags:
          type: array
          items:
            type: string
        importance:
          type: ["integer", "null"]
        memoryDate:
          type: ["string", "null"]
        createdAt:
          type: string
    AiMemoryInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        sourceType:
          type: string
          enum: [JOURNAL, HABIT, USER_NOTE, AI_SUMMARY, REFLECTION]
        tags:
          type: array
          items:
            type: string
        importance:
          type: integer
        memoryDate:
          type: string
    AiConversation:
      type: object
      required: [id, userId, title, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    AiConversationInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
    AiMessage:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdAt:
          type: string