# lib/api-zod/src/generated/types/aiMemoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Atomic Identity Habits API
 * OpenAPI spec version: 0.1.0
 */
import type { AiMemoryInputSourceType } from './aiMemoryInputSourceType';

export interface AiMemoryInput {
  content: string;
  sourceType?: AiMemoryInputSourceType;
  tags?: string[];
  importance?: number;
  memoryDate?: string;
}